biomass.raster.from.treelist.dt {slidaRtools}	R Documentation

Convert a list of trees to a above ground biomass raster

Description

Function that creates a raster of above ground biomass (AGB) from a list of trees. The raster can either represent crown-distributed or stem-localized AGB.

Usage

biomass.raster.from.treelist.dt(
  trees.dt,
  minx = 0,
  maxx,
  miny = 0,
  maxy,
  dist = "uniform",
  periodic = F,
  sd = 0.25
)

Arguments

trees.dt
A list of trees provided as a data.table with specific requirements: Each row represents one individual. The columns need to have the following names and content: X = X-coordinate of tree position, Y = Y-coordinate of tree position; CD = tree crown diameter (m), AGB = above ground biomass (t). Ideally the data should also have a column TreeID with unique IDs for each individual.

minx
Minimal X-coordinate

maxx
Maximal X-coordinate

miny
Minimal Y-coordinate

maxy
Maximal Y-coordinate

dist
Parameter specifying how AGB will be distributed in space: options are "stem" for stem-localized, "uniform" for uniformly crown-distributed or "Gaussian" for bell shaped crown-distributed.

periodic
Boolean enabling periodic boundaries, i.e. if set to TRUE crown parts of trees that exceed the edges of the area reappear on the opposite side.

sd
Standard deviation of the optional Gaussian distribution (default = 0.25)

Value

Raster object

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]