calc.FCC.from.point.cloud {slidaRtools}	R Documentation

Calculate fractional canopy cover (FCC) from XYZ-dataframe

Description

Function to calculate FCC from the point cloud in XYZ-dataframe format for a custom pixel size and height threshold. FCC is the relative share of all CHM pixels values above a height threshold. The CHM is generated internally from the point cloud by rasterizing the highest return in each pixel of a given resolution.

Usage

calc.FCC.from.point.cloud(XYZdf, pxl.res = 1, h.threshold = 1)

Arguments

XYZdf
Point cloud in XYZ-dataframe format

pxl.res
Side length of CHM pixel

h.threshold
Height threshold (pixels with the exact threshold value are included as canopy)

Value

Fractional canopy cover (FCC) as a value between 0 and 1

Examples

in progress

[Package slidaRtools version 0.1.0 Index]