calc.P.H.ratio {slidaRtools}	R Documentation

Calculate P:H ratio

Description

Function that calculates the ratio P:H of a canopy profile (Marvin et al. 2014), where P is the height of the maximal peak in the profile (highest return or foliage density) and H is the maximal height of the profile.

Usage

calc.P.H.ratio(prof)

Arguments

prof
Lidar profile as vector with height bins as names

Value

number P:H

Examples

in progress

[Package slidaRtools version 0.1.0 Index]