calc.ShannonIndex {slidaRtools}	R Documentation

Calculate the Shannon index of a profile

Description

Function that returns the Shannon diversity index of a vector of counts or relative proportions (e.g. Lidar profile).

Usage

calc.ShannonIndex(vec, ex.vec = NA)

Arguments

vec
Vector of counts or relative proportions

ex.vec
Vector indices of elements in vec that should be excluded (e.g. to exclude ground return of a Lidar profile set to 1)

Value

Shannon index

Examples

in progress

[Package slidaRtools version 0.1.0 Index]