calc.TCH.from.point.cloud {slidaRtools}	R Documentation

Calculate mean top-of-canopy height (TCH) from XYZ-dataframe

Description

Function to calculate TCH from the point cloud in XYZ-dataframe format for a custom pixel size. TCH is the mean of all CHM pixel values. So within each pixel the highest return is taken and then the average height over all pixels is calculated.

Usage

calc.TCH.from.point.cloud(XYZdf, pxl.res = 1)

Arguments

XYZdf
Point cloud in XYZ-dataframe format

pxl.res
Side length of CHM pixel

Value

mean top-of-canopy height (TCH)

Examples

in progress

[Package slidaRtools version 0.1.0 Index]