calc.cumulative.size.index {slidaRtools}	R Documentation

A function to calculate the CSI cummulative Size Index

Description

A function to calculate the CSI cummulative Size Index

Usage

calc.cumulative.size.index(dbh, h_max, Xfactor = 0.6, part2 = "old")

Arguments

dbh,
vector of all dbh of a plot in m

h_max,
height of the talles tree in the plot

Xfactor,
which Fraction of the cummulative dbh should be devided by the cummulative dbh

part2,
"no" returns only part 1, "dbh_max" multilpies part1 with dbhmax^2.5; "h_max" multiplies part 1 with f(h_max,dbh_max);"old" with h_max^1.5

Value

dimensionless CSI

Author(s)

Friedrich J. Bohn

Examples

in progress

[Package slidaRtools version 0.1.0 Index]