calc.lidar.vert.index.DLR2.rel.centr.mass.from.prof {slidaRtools}	R Documentation

Calculate vertical structure index suggestion of Victor Cazcarra (DLR) from Lidar profile or vertical foliage profile

Description

Function for vertical index from DLR calculated from Lidar XYZ-table data (point cloud or CHM). The equation for the index V is: V = 1 - (abs(CM - Hperc) / Hperc) with CM being the center of mass (equal to MCH or TCH, depending if input is point cloud or CHM-raster) and Hperc being a custom fraction of the max. height (Hmax) in the plot (default 50 percent).

Usage

calc.lidar.vert.index.DLR2.rel.centr.mass.from.prof(
  prof,
  perc = 0.5,
  h.min = 1
)

Arguments

perc
Fraction between 0 and 1 (default 0.5) to specify Hperc = perc * Hmax

h.min
Minimum height below which Lidar returns are ignored

XYZtable
Lidar point cloud or CHM in XYZ-table format

Value

Index of vertical forest structure

Examples

in progress

[Package slidaRtools version 0.1.0 Index]