display.point.cloud.dt {slidaRtools}	R Documentation

Plot point cloud in 3D space

Description

Function for 3-dimensional point cloud visualization.

Usage

display.point.cloud.dt(
  xyz.dt,
  col.palette = NA,
  col.var = "Z",
  col.lim = NA,
  size = 1
)

Arguments

col.palette
Color palette

col.var
String with name of the column which contains the information for coloring

col.lim
Vector of two elements specifying the limits of the col.var values

size
...of plotted points

pc
Point cloud in XYZ-table format (important: column names need to be X, Y and Z)

Value

3D plot in rgl window

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

[Package slidaRtools version 0.1.0 Index]