display.profile {slidaRtools}	R Documentation

Plot vertical profile, e.g. of a Lidar dataset

Description

Function for profile plotting.

Usage

display.profile(
  prof.vec,
  line.col = "black",
  fill.col = NA,
  add = F,
  maxx = NA,
  maxy = NA,
  alpha.line = 1,
  alpha.fill = 0.5,
  lwd = 2,
  lty = 1,
  GR.mark = NA,
  GR.mark.cex = 1,
  GR.mark.lwd = 2,
  GR.mark.col = NA,
  alpha.GR.mark = 1,
  xlab = bquote("Count of Lidar returns [ha"^-1 ~ "]"),
  ylab = "Height [m]",
  cex.axis = 1,
  cex.lab = 1,
  las = 1
)

Arguments

prof.vec
Profile with values and height bins as names (derived from make.profile functions)

line.col
Line color

fill.col
Polygon fill color (set to NA if no fill is desired)

add
If F a new graph is drawn, if T the profile is plotted into an existing graph

maxx
X-axis maximum

maxy
Y-axis maximum

alpha.line
Transparency setting for the line

alpha.fill
Transparency setting for the polygon fill

lwd
Line width

lty
Line type

GR.mark
Character that should be displayed at the tip of the ground return peak (if none set to NA)

GR.mark.cex
Size of the ground return marker

GR.mark.lwd
Line width of the ground return marker

GR.mark.col
Color of the ground return marker

alpha.GR.mark
Transparency setting for the ground return marker

xlab
X-axis titel

ylab
Y-axis titel

cex.axis
Font size of axis numbers

cex.lab
Font size of axis titels

las
Orientation of axis numbers

Value

Graphic of the profile

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]