extract.point.values.from.raster {slidaRtools}	R Documentation

Extract point values from raster

Description

Function that extracts values from an underlying raster at the position of each point of an overlying point cloud.

Usage

extract.point.values.from.raster(pc, ras)

Arguments

pc
Input point cloud

ras
Input raster

Value

Vector of raster values for each point, that can be added as a new column to the point cloud dataframe

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]