find.nearest.neighbors {slidaRtools}	R Documentation

Find nearest neighbors in point cloud

Description

Function that finds the n nearest neighbor points for each point in a point cloud of up to 3 dimensions

Usage

find.nearest.neighbors(pc, n = 2)

Arguments

pc
Point cloud as data.frame with at least a column "X" and optional columns "Y" and "Z"

n
Number of neighbors to find (1 only returns the focal point itself)

Value

list with up to five elements which are all matrices where each row represents each focal point in the point cloud and each column represents one neighbor sorted by increasing distance. The matrix elements represent 1) neighbor indices, 2) neighbor distances, 3) neighbor X-coodinates, 4) neighbor Y-coodinates, 5) neighbor Z-coodinates.

Examples

in progress

[Package slidaRtools version 0.1.0 Index]