in.cuboid {slidaRtools}	R Documentation

Check if point lies within a cuboid

Description

Function that checks for each X-Y-coordinate pair, whether the point lies in a cuboid of a given center, side length and height. Works with single numbers or matrices of X- and Y-coordinates as input.

Usage

in.cuboid(Xcor, Ycor, Zcor, Xctr, Yctr, Zbase, sidelength, height)

Arguments

Xcor
X-coordinate of the point

Ycor
Y-coordinate of the point

Zcor
Z-coordinate of the point

Xctr
X-coordinate of the center

Yctr
Y-coordinate of the center

Zbase
Z-coordinate of the lower end

sidelength
horizontal length and width of the cuboid

height
...of the cuboid

Value

boolean or matrix of booleans

Examples

in progress

[Package slidaRtools version 0.1.0 Index]