make.lidarscan.dt {slidaRtools}	R Documentation

Scan a voxel forest with a virtual Lidar

Description

Function that simulates a Lidar point cloud of a given voxel forest and writes it to a file in XYZ-table format. The model uses a Beer-Lambert light extinction approach and interprets it as probability to get a Lidar return from a voxel.

Usage

make.lidarscan.dt(
  vxf.dt,
  P0.AGR = 0.2,
  k.AGR = 0.2,
  P0.GR = 0.2,
  k.GR = 0.2,
  LAD.prefactor = F,
  prob.out = F
)

Arguments

vxf.dt
data.table containing the voxel forest (derived from make.voxelforest.dt function)

P0.AGR
Initial return probability for above ground voxels

k.AGR
Extinction coefficient for above ground voxels

P0.GR
Initial return probability for ground voxels

k.GR
Extinction coefficient for ground voxels

LAD.prefactor
Boolean to specify whether leaf area density (LAD) of a voxel should be multiplied with P0 to get final return probability or not

prob.out
Boolean to specify whether an additional output containing the full voxelforest with information on canopy above and Lidar return probability for each voxel should be produced

Value

data.table object containing a XYZ-table of discrete Lidar returns (point cloud). If prob.out=T the function will return a list containing two data.tables: 1) the Lidar point cloud and 2) the voxel forest with Lidar return probability info.

Examples

in progress

[Package slidaRtools version 0.1.0 Index]