make.map.from.spatial.index {slidaRtools}	R Documentation

Make rasters from values and spatial indices

Description

Function that returns a raster based on given spatial indices (plot numbers) and corresponding values that should be assigned to each pixel.

Usage

make.map.from.spatial.index(
  spatial.ID,
  value,
  res = 1,
  minx = NA,
  miny = NA,
  maxx = NA,
  maxy = NA
)

Arguments

spatial.ID
vector containing spatial indices

value
vector containing values to fill the pixels

res
side length of one spatial subunit

minx
minimum X-coordinate

miny
minimum Y-coordinate

maxx
maximum X-coordinate

maxy
maximum Y-coordinate

Value

raster object

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]