make.profile.from.XYZ.value {slidaRtools}	R Documentation

Derive a vertical profile of a point cloud for a certain variable

Description

Function that creates profiles of a certain statistic of a certain variable from XYZ-tables with a custom height resolution (binwidth)

Usage

make.profile.from.XYZ.value(XYZ.value.table, binwidth = 1, stat = "sum")

Arguments

XYZ.value.table
4-column data frame with 3rd column containing heights and 4th column containing values for which to make the profile

binwidth
number specifying the vertical width of height classes

stat
desired statistic ("sum", "count", "max", "min", "mean", "median")

Value

vector of point or pixel counts per height class with height values as names

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]