make.profile.from.Z {slidaRtools}	R Documentation

Derive a vertical profile of a vector of heights

Description

Function that creates profiles from Z-vectors with a custom height resolution (binwidth)

Usage

make.profile.from.Z(Z.vec, binwidth = 1)

Arguments

Z.vec
vector containing single height measurements

binwidth
number specifying the vertical width of height classes

Value

vector of point or pixel counts per height class with height values as names

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]