periodic.boundaries.dt {slidaRtools}	R Documentation

Periodic boundaries

Description

Function that creates periodic boundaries for point clouds, i.e. points that exceed the boundaries of the area re-appear on the opposite side of the area.

Usage

periodic.boundaries.dt(xyz.dt, minx, maxx, miny, maxy)

Arguments

xyz.dt
Data.table containing X-, Y- and Z-coordinates of the points

minx
Minimum X-coordinate of the area

maxx
Maximum X-coordinate of the area

miny
Minimum Y-coordinate of the area

maxy
Maximum Y-coordinate of the area

Value

data.table of X-, Y- and Z-coordinates with periodic boundaries

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]