periodic.boundaries.for.circular.plot {slidaRtools}	R Documentation

Periodic boundaries for circular forest plots

Description

Function that creates periodic boundaries for voxelforests along circular plot borders (e.g. large lidar footprints or circular inventory plots). Tree crown voxels that exceed the boundaries of the circle re-appear on the opposite side of the circle. The placement of trees is based on the polar coordinates of the stem positions (distance and angle to the circle center). All trees with stems inside the circle are copied to exist a second time outside of the circle at the position opposite of the circle center and with the same distance to the boundary, but outwards.

Usage

periodic.boundaries.for.circular.plot(vxf.dt, Xctr, Yctr, radius, clip = T)

Arguments

vxf.dt
Data.table containing voxelforest object created with make.voxelforest; columns of the stem positions Xstem and Ystem are crucial

Xctr
X-coordinate of the plot center

Yctr
Y-coordinate of the plot center

radius
of the circular plot

clip
Boolean to specify whether output should be clipped to the circle (default) or all

Value

data.table containing the voxelforest with periodic boundaries

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]