periodic.raster.frame {slidaRtools}	R Documentation

Periodic raster frame

Description

Function that increases a raster by adding a frame around in which the pixel values correspond to the values found at the opposite side of the raster, creating periodic boundaries, e.g. for moving window analysis.

Usage

periodic.raster.frame(ras, w = 10)

Arguments

ras
Raster object

w
Width of the frame (number of pixels)

Value

raster object

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]