ras2colmx {slidaRtools}	R Documentation

Convert raster to bycol-matrix

Description

Convert a raster with origin in lower left corner to a matrix with origin in upper left corner and with cell order filling column after column from top to bottom (bycol)

Usage

ras2colmx(ras)

Arguments

ras
Input raster object

Value

Matrix (cell order bycol)

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]