raster.from.point.cloud {slidaRtools}	R Documentation

Convert a point cloud to a raster

Description

Function that creates a raster from a point cloud by applying a certain function in each cell, e.g. func=max returns a surface raster (CHM in case of vegetation), min returns lowest point per cell, mean returns the mean value of all points in the cell, count returns a point density raster, sum returns the sum of all point values and mode returns the most common value (in case of several equally common values one of them is taken randomly).

Usage

raster.from.point.cloud(
  pc,
  xcor = "X",
  ycor = "Y",
  var = "Z",
  res = 1,
  func = "max"
)

Arguments

pc
Point cloud in XYZ-table format

xcor
Variable name that contains the X-coordinates (Default "X")

ycor
Variable name that contains the Y-coordinates (Default "Y")

var
Variable name that contains the values that should be rasterized (Default "Z")

res
Resolution (cell side length) of the output raster

func
Function to be applied (either "max", "min", "mean", "count", "sum" or "mode")

Value

Raster object

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]