rbind.large.dt.files {slidaRtools}	R Documentation

Read large tables from files and rbind them into one huge data table

Description

Function that reads a group of equally structured tables from files (e.g. point clouds in XYZ-format) and integrates them into one huge data table

Usage

## S3 method for class 'large.dt.files'
rbind(path, pattern)

Arguments

path
...to the directory with the input files

pattern
...that is containded in all input filenames

Value

data table containing all rows of all input tables

Examples

in progress

[Package slidaRtools version 0.1.0 Index]