remove.outliers.from.point.cloud.IQR {slidaRtools}	R Documentation

Remove outliers from point cloud

Description

Function that removes points from a point cloud if their Z-value exceeds the given percentile of their neighborhood's Z-values by a certain factor.

Usage

remove.outliers.from.point.cloud.IQR(
  pc,
  n = 50,
  low.perc = 0.25,
  high.perc = 0.75,
  fac = 1.5,
  high = T,
  low = T
)

Arguments

pc
Input point cloud (first 3 columns need to be "X", "Y" and "Z")

n
Number of considered neighbor points (with lowest XY-distance to focal point)

low.perc
Lower end of the desired interquantile range (fraction between 0 and 1)

high.perc
Upper end of the desired interquantile range (fraction between 0 and 1)

fac
Factor by which a Z-value needs to exceed or subceed the IQR of its neighbors to be considered an outlier

high
bool if false high outliers will not be removed

low
bool if false low outliers will not be removed

Value

Point cloud dataframe with rows from the input that were detected as outliers removed

Examples

in progress

[Package slidaRtools version 0.1.0 Index]