remove.outliers.from.raster.IQR {slidaRtools}	R Documentation

Remove outliers from raster

Description

Function that sets value of a raster pixel to NA if it exceeds or subceeds the interquantile range (IQR) of its neighbors by a certain factor.

Usage

remove.outliers.from.raster.IQR(
  ras,
  rad = 5,
  low.perc = 0.25,
  high.perc = 0.75,
  fac = 1.5,
  high = T,
  low = T
)

Arguments

ras
Input raster

rad
Radius of the considered (circular) neighborhood

low.perc
Lower end of the desired interquantile range (fraction between 0 and 1)

high.perc
Upper end of the desired interquantile range (fraction between 0 and 1)

fac
Factor by which a pixel value needs to exceed or subceed the IQR of its neighbors to be considered an outlier

high
bool if false high outliers will not be removed

low
bool if false low outliers will not be removed

Value

Raster object with gaps (NA) at the positions of former outliers

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]