rotate.points {slidaRtools}	R Documentation

Rotate points around X-, Y- or Z-axis

Description

Function to rotate a given set of points or point cloud in a plain around center point with a certain angle.

Usage

rotate.points(cor.df, rot.angle, rot.axis = "Z", center = c(0, 0, 0))

Arguments

cor.df
Coordinate dataframe with columns containing X-, Y- and Z-coordinate (at least two columns are required and names need to be "X", "Y" or "Z", respectively)

rot.angle
Rotation angle in degrees (clockwise rotation)

rot.axis
Rotation axis (default is "Z", can also be set to "X" or "Y")

center
Vector of three elements specifying the X-, Y- and Z-coordinate of the rotation center point

Value

Coordinate dataframe with X- and Y-coordinates after rotation

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]