rowmx2ras {slidaRtools}	R Documentation

Convert byrow-matrix to raster

Description

Convert a matrix with origin in upper left corner and with cell order from left to right (byrow) to a raster with origin in lower left corner

Usage

rowmx2ras(mx)

Arguments

mx
Input matrix

Value

Raster object

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]