sobol.tree.positions {slidaRtools}	R Documentation

Distribute trees evenly in the plot considering their sizes

Description

Function that distributes trees evenly (quasi randomly) in the plot starting with the largest and positioning them one after another with decreasing size using the Sobol algorithm. The sobol algorithm needs initialization before the first application of the function (see example below).

Usage

sobol.tree.positions(treesize.vec, plotsize = 20)

Arguments

treesize.vec
Vector of tree sizes (e.g. DBH, BA or height)

plotsize
Sidelength of a square shaped plot [m]

Value

Data.frame with 3 columns containing X- and Y-coordinates and tree sizes

Author(s)

Nikolai Knapp

Examples

sobol.init <- sobol(1, dim=2, init=T)
pos <- sobol.tree.positions(c(3, 2, 1), plotsize=10)
plot(pos$Y ~ pos$X)

[Package slidaRtools version 0.1.0 Index]