terrain.under.voxelforest {slidaRtools}	R Documentation

Add terrain under a voxelforest

Description

Function that shifts a voxelforest vertically to add underlying terrain elevation.

Usage

terrain.under.voxelforest(vxf.dt, terrain.dt)

Arguments

vxf.dt
Input data.table which contains a voxel forest created with make.voxelforest function and the specific settings "keep=c("TreeID", "X", "Y")", i.e. each voxel has to contain the unique tree ID and the stem position

terrain.dt
Input data.table which contains X-, Y- and Z-column representing a digital terrain model at 1-m resolution (e.g., can be obtained from a 1-m DTM raster with data.table(ras2xyzdf(dtm.ras)))

Value

Data.table containing a voxel forest with terrain

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

[Package slidaRtools version 0.1.0 Index]