thin.point.cloud {slidaRtools}	R Documentation

Point cloud thinning

Description

Function that thins a point cloud to a desired homogenous horizontal point density by keeping a random subset of points of a desired size in each spatial cell and discarding all other points.

Usage

thin.point.cloud(pc, cell.size = 1, returns.per.cell = 4)

Arguments

pc
Point cloud in XYZ-table format (important: column names need to be X, Y and Z)

cell.size
Side length of a spatial unit

returns.per.cell
Number of returns that should be kept in each cell

Value

data.frame of the thinned point cloud in XYZ-table format

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

in progress

[Package slidaRtools version 0.1.0 Index]