vertical.foliage.profile {slidaRtools}	R Documentation

Derive a vertical foliage profile (VFP) from a Lidar profile

Description

Function that returns the vertical foliage profile for a given Lidar profile using the approach of Harding et al. (2001) which is based on the MacArthur-Horn method.

Usage

vertical.foliage.profile(profile, h.bin = 1, GR.threshold = 5, k = 0.5)

Arguments

profile
Vector of Lidar profile (with heights as names)

h.bin
Vertical extent of one height bin

GR.threshold
Height threshold below which everything is regarded as ground return (GR) and ignored in the processing

k
Light extinction coefficient, which is an adjustment factor from relative to absolute (LAD) VFP

Value

Vector of the vertical foliage profile (VFP)

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

Examples

prof.vec <- c(9, 2, 4, 5, 6, 3, 1)
names(prof.vec) <- 0:6
vfp.vec <- vertical.foliage.profile(profile=prof.vec, h.bin=1, GR.threshold=1, k=0.5)
plot(names(prof.vec) ~ prof.vec, type="l", xlab="Lidar returns", ylab="Height [m]")
plot(names(vfp.vec) ~ vfp.vec, type="l", xlab="LAD [m2/m3]", ylab="Height [m]")

[Package slidaRtools version 0.1.0 Index]