weighted.returns.point.cloud {slidaRtools}	R Documentation

Assign a weight to each point in a point cloud depending on total count points in voxel column

Description

Function that normalizes return weights in point clouds by dividing the value 1 by the total count of returns in a voxel column of a given horizontal resolution, e.g. if there are 4 returns in one vertical column each gets a weight of 0.25

Usage

weighted.returns.point.cloud(pc, cell.size = 1)

Arguments

pc
Data frame of the point cloud with first 3 columns representing X-, Y- and Z-coordinates

cell.size
Horizontal side length of one spatial subunit (voxel column width in X- and Y-direction)

Value

Data frame of the point cloud with additional column "ReturnWeight"

Examples

in progress

[Package slidaRtools version 0.1.0 Index]